add abstract class CMarioForm

+++ Final_Project/CBaseCharacter.h
@@ -14,7 +14,7 @@ protected:
 	float accelX;
 	float accelY;
 
-	float dirX;
+	int dirX;
 
 public:
 };
